# go/echo/api-gateway/service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Kyunghwan Kim",
            "email": "khsyee@gmail.com"
        }
    },
    "paths": {
        "/": {
            "get": {
                "description": "It returns true if the api server is alive.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Healthcheck",
                "responses": {
                    "200": {
                        "description": "API server's liveness",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/file": {
            "post": {
                "description": "Send file",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Send file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/sum": {
            "post": {
                "description": "It outputs sum of two inputs",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sum integer inputs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "inputX",
                        "name": "inputX",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "inputY",
                        "name": "inputY",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Integer response: int",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}